requirements.txt
fastapi==0.104.1
gunicorn==23.0.0
matplotlib==3.8.0
numpy==1.26.0
opencv-python-headless==4.8.1.78 
Pillow==10.0.1
python-dotenv==1.1.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.3
uvicorn==0.23.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.0+cpu
torchvision==0.18.0+cpu